[#1755] slipstream: T5.6 review hardening — non-fatal per-range enhancement, per-pass counter reset
Review follow-up on 3d7bf613:

1. Per-range interleaved enhancement (F3) is now NON-FATAL: it is an
   optimization (progressive tx visibility), not a correctness guarantee —
   that is the final post-loop run's job. The FFI runner is single-shot
   (sync_once error → SyncState::Error until app-lifecycle retry), so a
   transient gRPC blip at a range boundary must not abort a 16-minute
   restore. warn + continue instead of `?`.

2. Progress::begin_pass() called at sync_once entry: resets the per-pass
   RATIO counters (scanned, fetched, pass_total, range_end, spendable
   latch). The handle outlives passes (Swift prepare() opens once;
   stop()/start() reuse it across background/foreground), so a resumed
   restore previously computed scanned(stale)/pass_total(remaining) —
   59% at pass start, pinned at clamp later. Latent since T5.5; surfaced
   by F1's store semantics. DELTA counters (enhanced_txs,
   ranges_completed, reorgs_recovered) stay monotonic per handle — Swift
   consumes them via strict-greater last-seen comparisons.

3. Clippy clean (both feature sets): spendable latch collapsed to
   let-chain; SyncReport struct-init in test helper; MAX_CONSECUTIVE_REORGS
   bounds moved to compile-time const assertions; stray `u8 as u8`.

Also restored the canonical 3-slice LocalPackages XCFramework (the T5.6
build left a nested xcframework with only a macOS slice at top level).

Tests: cargo 64+14/0, clippy 0/0 (default + darkside), darkside 6/6
(incl. reorg_recovery_produces_correct_tip through the new per-range
enhancement scheduler path), OfflineTests 463/0.

+++ slipstream/core/src/verify.rs
@@ -66,7 +66,7 @@ mod tests {
         c.verify_blocks(&all[..5]).expect("first half");
         // Second half must link to the first (its first prev_hash = hash of 104).
         let mut second = all[5..].to_vec();
-        second[0].prev_hash = vec![104u8 as u8; 32];
+        second[0].prev_hash = vec![104u8; 32];
         c.verify_blocks(&second).expect("second half");
         assert_eq!(c.last_height, Some(109));
     }
